Sets up initial layout.
This is the initial plans for the layout - a control bar at the top
to open a menu to configure the timer, then two timers (one for the
half and one for the jam) and two buttons (one for timeouts and one
for jam start/end). May still play around with alignment, and need to
figure out labeling for the two timers. Dimensions are guessing at
2:4.5 h:w ratio (4 numbers 2:1, and a colon at half width).