Adds jam timer functionality to Timer component
This commit modifies the Timer component so that it has multiple
useTimer objects for the various timer functions that the WFTDA Jam
Timer NSO measures, as well as functions and logic to recognize the
appropriate game state and flow into the following game states. It
also makes use of the useEffect hook to recognize when either the jam
time or lineup time has expired and automatically trigger game
state progression.

This commit also makes updates to the useTimer hook to memoize the
startTimer and resetTimer functions via the useCallback hook, in order
to pass the eslint excaustive-deps linting warning in the useEffect
hook described in the paragraph above. This also required a bit of
refactoring the useTimer hook, so that instead of one 'status' string
state variable, it uses separate boolean 'paused' and 'limit' state
variables.