Refactors time display into its own component
This commit moves the rendering of the timer displays from the Timer
component into a new Display component. The display component takes
props for the time, whether the timer is countdown or count-up, and
the max time (only needed for countdown). It also takes a style prop.

The Timer/index component also has some additional new commented out
code. I am working on another refactoring of this component, but
decided it would be better to pull out the Display function into its
own component first.