Moves button functions into their own component
This commit moves the functionality for the buttons used to control
the timers out of the Timer component into a new ControlButton
component. The component takes as props the label for the button and
the onPress function to assign to it. It also takes a style prop for
the text display.