Modifies useTimer/pauseTimer for memoization
This commit modifies the pauseTimer function in the useTimer hook so
that it does not rely on the time parameter (which is updating every
10ms), but instead calculates the elapsed time at the time pause is
called. This is done so that the function can be memoized with the
useCallback hook without including the time parameter in the array of
dependencies (passing the eslint exhaustive-deps warning), so the
memoized function will not be updated every 10ms.