Refactors useTimer and Timer component for sync
When testing the previous commit, I discovered that if the timers
were allowed to run to their limits the two clocks would slowly get
out of sync due to the processing time of the various starts, stops,
and resets. This is probably less than one would experience with a
human jam timer holding two stopwatches, but it was still jarring to
observe.

This commit makes changes to avoid the loss of synchronization
between the timers. It moves the functionality of updating the time
out of the useTimer hook and into a useEffect in the Timer component
- the hooks now simply track the start, pause, and max time for the
timer, and whether it is running, paused, or has reached its max
time, while the Component tracks the time to display and
simultaneously updates both displays every 10ms.

This commit also introduces a combined RESETANDSTART dispatch for the
useTimer, which both resets the timer and immediately starts it
running - this eliminates delay between the two actions when they are
intended to be simultaneous.

This commit further updates several of the dispatch commands to
optionally take a "currTime" parameter in the payload, which is the
system time when the command was intended to be issued. This is
either measured as soon as a button is pressed, or calculated to when
the maxTime should have happened.