Adds timer configuration
This commit adds the ability to change the max time for each of the
timers in the Configuration view. It adds the ConfigEntry component
which is a form that lets the user update the minutes and seconds for
each timer.

The ConfigEntry component is not yet formatted to any usable degree -
right now it is simply functional to test that it can work.