Improves formatting for nav buttons and config
This commit makes changes to the nav menu buttons and config screen for
better UI.

Nav Buttons are changed from a React Router Native Link element inside a
React Native Pressable to a useNavigate function in the onPress
attribute of the Pressable. This makes it so the entire button is active
(rather than the padding not being active) and fixes the color styles to
work as intended.

The config screen now sets the width of the Textbox components to a
fixed width. This ensures that they do not shrink as a user is entering
a number when fewer than two digits are present.