Adds alerts when second timer is close to limit
This commit adds a visual alert by flashing the background of the second
timer red when the timer is within 5 seconds of its limit. This is
currently hardcoded as applying only to the second timer, and the 5
seconds is hardcoded but made a part of the ConfigContext so it can be
potentially adjusted in app settings in future revisions.